go-bindata: balance backquoted string concatenation
The current code prints long string concatenations like

	a + b + c + d + e + f

The new code prints

	((a + (b + c)) + (d + (e + f)))

This has the effect of minimizing the overall height of the
expression tree, which makes recursive traversals of the
tree a little happier and makes incremental computation
of string constants as in go/types and go/vet MUCH happier.

The compilers and go/types should be fixed so that this
isn't necessary, but we can't fix the old releases, and this
is very easy, so it seems worth doing.

For golang/go#23222.